requirements.txt
torch
matplotlib
numpy
pyyaml